+++ relayRun/src/main/java/com/example/relayRun/config/SecurityConfig.java
@@ -3,8 +3,8 @@ package com.example.relayRun.config;
 import com.example.relayRun.jwt.TokenProvider;
 import com.example.relayRun.jwt.exception.JwtAccessDeniedHandler;
 import com.example.relayRun.jwt.exception.JwtAuthenticationEntryPoint;
-import com.example.relayRun.user.oauth2.OAuth2SuccessHandler;
-import com.example.relayRun.user.service.CustomOAuth2UserService;
+import com.example.relayRun.oauth2.controller.OAuth2SuccessHandler;
+import com.example.relayRun.oauth2.service.CustomOAuth2UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
