+++ relayRun/src/main/java/com/example/relayRun/oauth2/controller/OAuth2Controller.java
@@ -1,4 +1,4 @@
-package com.example.relayRun.user.oauth2;
+package com.example.relayRun.oauth2.controller;
 
 import org.springframework.web.bind.annotation.*;
