--- relayRun/src/main/java/com/example/relayRun/user/oauth2/OAuth2Attribute.java
@@ -1,16 +1,0 @@
-package com.example.relayRun.user.oauth2;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.Map;
-
-@Getter
-@Builder
-public class OAuth2Attribute {
-    private Map<String, Object> attributes;
-    private String name;
-    private String email;
-    private String imgURL;
-
-}
